Add %ask/%tell to the cog interface.
[%cog %send/%recv ...] only work in a single direction, meaning you can't
implement more complicated semantics like an implementation of the STM
interface on top of it. You get an acknowledgment that the remote message was
delivered, but you can't actually implement atomic actions on top of it.

So we add [%cog %ask/%tell ...] instead. %ask sends a value to a %tell, but
%tell has a function in its request which the %ask value is run against, and a
value is given in response to both the %ask and the %tell.

We add two test programs to exercise this:

- sire/demo_cog_asktell_basic.sire: tests four callers against a single tell
  interface as a test. The %tell cog just hands out monotonically increasing
  numbers.

- sire/demo_cog_asktell_askcrash.sire: tests that when the %tell function
  crashes, it crashes the %tell cog instead of passing a crash value into
  the %ask-ing cog.