sire: remove test.sire
`test.sire` was a file that depended on all of the test files, so that
you could run all the tests with `plunder sire sire/test.sire`.

However sh/sire-load-all obliviates then need for this, since it
efficiently tests all sire files (including all of those test files).

This also opens the door to making each test file *not* all live in a
single load-order sequence.