rts: implement PLANREC
This switches the runtime over to supporting the new (and hopefully
final) version of PLAN, which supports recursive bindings.

This *removes* support for lets-within-apps, but the current version of
Sire has already been updated to not generate code like that.

Because the everything-is-LETREC change makes things measurably slower,
this patch also includes an optimization pass that breaks up big LETRECs
into a string of LETs and smaller LETRECs, and then floats bindings down
the tree, as close to their usage-sites as possible.

Also, because this patch only creates `LET` nodes for bindings that have
no self-reference, the interpreter can omit the self-reference machinery
for most bindings.

Altogether, These changes end up causing to substantial *speedup*, which
is a pleasant surprise.