rts: all _JetCandidates in the jets table
This adds all "jet candidates" (functions named using the _JetNaming
convention) to the jets table, but does so explicitly without
implementation.

The primary motivation here is to just have the jets table match the
stderr output for jet hashes.  This is especially important in CI, where
we want to error out if there is a mismatch between the Sire code and
the runtime's expectations of it.

A nice side-effect is that there is an explicit list of "unimplemented
jets" in `Fan.JetImpl`.