sire: _BarSliceToNat (with jet)
This jets the trivial `(barNat (barSlice ...)`

This may seem like a pointless jet, but generally the
"_BarSlice" variants are the most important bar jets, since
working with slices will be very common in practice, and
the slice-oriented jets are strictly more general than their
Bar counter-parts.

Beyond that, this is particular jet very important function for
treating an array as a word-buffer.
`(_BarSliceToNat off*wid wid)` is equivalent to indexing into
a packed array.