rts: explicit --save for plunder repl
At the moment, the output from the `trace` jet goes to stdout.
Before this patch, this output was intermixed with the formal
output given (as a list of bars) by the actual PLAN code in
the seed file.

Moving the `trace` output to stderr is problematic because
that gets mixed with the debug output.  Worse, the rex ci-tests
capture `trace` output, so mixing it up with other logging
breaks that test.

This patch solves the problem by providing an explicit --save output
which allows the "formal PLAN output" to be written to a file.

This will be used later on in this patchset, to perform
implement Sire quine test in CI.