sire,rts: rework rex repr, remove datajet
This eliminate the REX datajet and switches to a simpler PLAN encoding
of Rex trees.

Sire now stores Rex trees in Haskell as PlanRex, which is an encoding of
Rex that tolerates malformed inputs, and which remembers the underlying
noun.  This is important because we don't want to need to do a full
load/store every time we inspect Rex from Haskell.  This is the main
innovation that makes it possible to eliminate the datajet.

Loot still prints rex trees as Rex, which is possible because the new
encoding of Rex trees if peculiar.  This approach also improves the
printer because it now handles malformed Rex more gracefully.

This is also faster, because working with Rex doesn't require the Law
wrapper, and the data-jet machinery isn't triggered by each new Rex node.

An incidental changes needed to be included: A new debug-output path
needed to be created to replace the `trkM $ REX $ ...` path.