sire-in-sire: simplify error messages
Instead of attempting to pretty-print error messages,
sire-in-sire just outputs the type of error, and the sub-tree
that triggered the error.

The sire-in-sire quine is meant to be small and correct, and not
necessarily super helpful.  The Haskell version can be used to get
nicer errors.  Also, another variant of sire-in-sire can be produced
that is more usable.

Sire-in-sire needs to be super small (and fast) so that it can be used
by people that are creating their own runtime systems.