sire-in-sire: optimize by removing Either
Before this patch, the sire-in-sire front-end essentially worked in a
monad with this type:

    Either (Str, Rex) (SireState, Rex)

However, since the left-hand-size of the tuple is always a Nat in the bad
case, and never a nat in the happy case, we can eliminate the `Either`
wrapper altogether, and just switch on the state at each step.

This cuts the runtime of the Sire quine in half.