sire: stew compiles #switch manually
This patch updates Stew to stop using `switchE` to compile switch
statements to Rex, and instead to compile switch statements by
translating them to a lower-level Stew AST and then to re-invoke
the Stew compiler on the result.

This is a part of the broader push for Stew to handle all
code-generation itself, instead of re-using the macro-expansions
utilities defined during the bootstrapping process.