sire: allow macros to handle leaf-juxtapose
Previously leaf-juxtaposition was not accepted as input in Sire:

    foo{bar}

This patch adds a special case, that allows this syntax to be implemented
by giving to the `#` macro:

    foo{bar} -> foo#{bar}

This allows slightly nicer syntax for bytestrings:

    b"data", b{data}

And it also allows this type of prefix-syntax to be used to disambiguate
rows/tuples.

    row[3 4 5] -> row#[3 4 5]

    r[3 4 5] -> r#[3 4 5]

In Stew, we will need separate syntax for (rows vs tuples) and (records
vs tabs) because they have different types.  Using a rune to disambiguate
is kinda ugly because it needs paren-wrapping in many cases:

    #[x=(#[a=1 b=2])]

Writing without the inner parens doesn't work because the = and # runes
get globbed together:

    #[x=#[a=1 b=2]]

However, the name-prefix doesn't have this problem, and binds very
tightly:

    tab[x=tab[a=1 b=2]]

    t[x=t[a=1 b=2]]

Stew can mostly implement this behavior without the special-case added
by this patch, since the Stew parser can just implement that directly:

    (tabIdx {a} t[a=1 b=2])

But it can't implement that in the case of REPL input where that input
is given directly:

    t[a=1]

Because the Stew compiler would never be invoked (because the stew
compiler isn't bound to any macro that will handle this case).

With this patch, the stew compiler *can* handle this case, by binding
the {#} macro to the stew compiler.