cog: Add [%cog %wait cogid] to wait until a cog isn't running.
This adds a syscall to allow cogs to be notified once the cog they're
monitoring is no longer processing requests. Previously, the only way
to know this was to have a %reap open, but multiple cogs shouldn't have
%reaps because you don't know which one will get the state.

Here, there can be as many waiters as necessary, and they won't wait
for a possible %reap to be processed: this is about liveliness.