sire: add _Ifz jet
This implements the `_Ifz` jet (but does not jet it).  This is a useful
operation just because it can be implemented extremely efficiently in
a native runtime system.

Also, there's a lot of cases where `(if (isZero x) ...)` is used in code
instead of `(ifNot x)` because `if` is an operation on booleans and it
first casts the input to that before switching on it.  For example,

    (ifNot (1 2) %t %f) ==> %t

    (ifz (1 2) %t %f) ==> %f