sire: properties are owned by binds
Previously, sire properties were this database of metadata *about* binds,
and that database is a top-level slot in the sire state.

With that system, you could add new properties to old bindings, and this
was used for things like backfilling type information.

This caused problems with caching, because the total set of information
about a module wasn't actually frozen once the module was frozen.

This change reworks all usage-patterns of properties, so that the
properties are frozen with their binding, and live in the BIND tuple.

For example, instead of backfilling type information on existing binds,
we create new binds that have more properties.