sire: use BST tables in properties
This replaces all use of tabs with BST tables in binding properties.
(For example, type information, pattern matching information)

This is the first step towards removing sets and tabs from the bootstrap
process, which will significantly reduce its size and complexity.

Also, once this migration is complete, it will be much easier to get
a new runtime up to speed, because the bootstrapping process no longer
depends on the jets for tabs and sets.